refactor(backend): decompose admin_endpoints.py into sub-package
Splits the 1,164-LOC flat admin_endpoints.py into a sub-package:

  api/admin/
    __init__.py                     (21)  router re-export
    endpoints.py                    (42)  aggregator: includes sub-routers
    schemas.py                      (61)  Pydantic request/response models
    _common.py                      (54)  shared response-builder helpers
    queries.py                     (355)  raw SQL helper functions
    task_state.py                  (289)  cache-backed SyncTaskStore (prev commit)
    sync_service.py                (183)  background-task orchestration
    status_routes.py               (141)  /status, /statistics, /reference/status
    sync_publications_routes.py    (131)  POST/GET /sync/publications[/status]
    sync_variants_routes.py        (161)  POST/GET /sync/variants[/status]
    sync_reference_routes.py       (185)  /sync/reference/init, /sync/genes[/status]

Every file is under 500 LOC (max 355). HTTP surface is byte-identical —
verified by tests/test_http_surface_baseline.py locking in /admin/status
and by the 14 pre-existing admin tests in tests/test_admin_sync_endpoints.py
all passing unchanged.

Behavior changes:

1. The process-local _sync_tasks dict is gone. All sync task state now
   persists through SyncTaskStore → CacheService → Redis-or-fallback.
   Closes P3 #11 from the 2026-04-09 review ("in-memory task state is
   invisible across multi-worker deployments"). The store is keyed
   under admin:sync_task:{task_id} with a 24h TTL; the /status
   endpoints that fall back to "latest of this kind" use a per-kind
   latest pointer instead of scanning prefix-matched dict keys.

2. Task ids now have a uuid suffix (pub_sync_20260411_000000_abcdef01)
   so sub-second concurrent starts can never collide. Legacy ids
   without suffix simply return 404 from the /status endpoints after
   the 24h TTL expires — this is fine because there are no long-lived
   tasks in the old code either.

Import path update in main.py:
    from app.api import admin_endpoints, auth_endpoints
    from app.api import auth_endpoints
    from app.api.admin import router as admin_router
    app.include_router(admin_endpoints.router)
    app.include_router(admin_router)

Closes P2 #5 (admin_endpoints.py 1,159 LOC) and P3 #11 (_sync_tasks)
from docs/reviews/codebase-best-practices-review-2026-04-09.md.

Counts unchanged: 907 passed, 9 skipped, 3 xfailed.
