+++ backend/app/api/admin/__init__.py
@@ -1,8 +1,21 @@
 """Admin API sub-package.
 
 Split from the old flat ``app/api/admin_endpoints.py`` during Wave 4.
-The ``endpoints``/``schemas``/``queries``/``sync_service`` submodules
-land in a follow-up commit; the first commit introduces only the
-cache-backed ``task_state`` module so the new infrastructure can be
-reviewed (and tested) independently of the router split.
+Organised as:
+
+- ``schemas``       — Pydantic request/response models
+- ``queries``       — raw SQL helper queries (DB statistics, sync views)
+- ``task_state``    — cache-backed sync task state (replaces the old
+  in-memory ``_sync_tasks`` dict flagged in the 2026-04-09 review as
+  unsafe for multi-worker deployments)
+- ``sync_service``  — background-task orchestration for pub/variant/ref syncs
+- ``endpoints``     — thin FastAPI router delegating to the services above
+
+Public re-export::
+
+    from app.api.admin import router
 """
+
+from .endpoints import router
+
+__all__ = ["router"]
